# Yet another labels-to-Child lookup optimization