@njhill 2019-02-24
@brian-brazil conceptually the implementation is really quite simple, just happens to be tailored for the specific requirements here including not having to create a new object just for the purpose of hash/equals comparison, and not caring about modification performance relative to lookups. The java SDK includes various data structure implementations (such as `ConcurrentHashMap`), but these are aimed at broad variety of use cases / workloads, both in terms of the selection of classes included and in terms of how those classes are implemented.

For this reason there's a wealth of third-party utility libraries which fill in gaps and cater to more specific use cases for example the excellent [eclipse collections](https://github.com/eclipse/eclipse-collections) and [JCTools](https://github.com/JCTools/JCTools) libraries.

Impl-wise what's done here is kind of a combination of the JDK's `IdentityHashMap` and `CopyOnWriteArrayList`. I wrote it quite fast and can push an update to structure in a more readable way. It's also possible to de-duplicate much of the similar logic in the various `labels(...)` methods, but not so easy to do so without sacrificing some perf, so I'm still looking at that.

Given some of the reasons metrics are collected in the first place, I expect it's commonly done in performance sensitive contexts. Hence imho a first-order goal of a client library such as this should be minimal runtime overhead. This change eliminates all garbage on individual events and improves throughput by 2-3x.

Maintenance-wise I really don't feel it would be much of an issue given the simplicity of the data structure and the fact that this project is open source with many eyes on it.