@franz1981 2019-02-25
@brian-brazil I can close my https://github.com/prometheus/client_java/pull/445 to favour this one from @njhill :+1: 
I think his approach is much better!! 
And about
> For this reason there's a wealth of third-party utility libraries which fill in gaps and cater to more specific use cases for example the excellent [eclipse collections](https://github.com/eclipse/eclipse-collections) and [JCTools](https://github.com/JCTools/JCTools) libraries.

I vote for JCtools but I'm just biased on it :)