@brian-brazil 2019-02-25
> For this reason there's a wealth of third-party utility libraries which fill in gaps and cater to more specific use cases for example the excellent eclipse collections and JCTools libraries.

Dependencies are purposefully avoided in this library, to avoid issues including it.

> Maintenance-wise I really don't feel it would be much of an issue given the simplicity of the data structure and the fact that this project is open source with many eyes on it.

That doesn't mean that the implementation is correct. I prefer the other proposed approaches.