@njhill 2019-02-26
Thanks @brian-brazil @franz1981 

> Dependencies are purposefully avoided in this library, to avoid issues including it.

This makes a lot of sense, although if there was one which would be really useful you could still consider shading? But I wasn't actually suggesting any of those would necessarily be applicable here, just giving examples of how there are many specialized impls for different use cases which would never be added to the JDK itself.

> That doesn't mean that the implementation is correct.

Could you elaborate on what you mean by correct here? I did actually find a small bug and have pushed another commit which fixes that, has some other refinements, and adds a bunch of comments explaining that "big" method. Given your comments, I was still thinking to restructure it a bit further so that the "map" impl is better encapsulated/separated from the existing logic.

> I prefer the other proposed approaches.

I'd be interested to understand the reasoning behind this. This PR isn't much different to those in terms of amount of additional code. If you don't care about performance then why would you make any change at all? If you do care about performance then why would you not go with this approach that outperforms the others considerably (and eliminates all allocations on the hot path)?

FWIW here's new benchmarks on the latest update. I think the areformentioned hash "fix" may have actually improved things even more (note I ran on a machine with less noise this time so the baselines are higher too).

Before:
```
Benchmark     Mode  Cnt         Score        Error  Units
baseline     thrpt   20  91140429.290 ± 608950.114  ops/s
oneLabel     thrpt   20  39409899.908 ± 836615.183  ops/s
twoLabels    thrpt   20  35696861.007 ± 480466.417  ops/s
threeLabels  thrpt   20  28408385.256 ± 461281.284  ops/s
fourLabels   thrpt   20  26435543.820 ± 193723.507  ops/s
fiveLabels   thrpt   20  24995175.241 ± 996512.887  ops/s
```

After:
```
Benchmark     Mode  Cnt         Score        Error  Units
baseline     thrpt   20  91230169.066 ± 519996.338  ops/s
oneLabel     thrpt   20  90727727.142 ± 765178.475  ops/s
twoLabels    thrpt   20  90294849.112 ± 939603.445  ops/s
threeLabels  thrpt   20  81616654.625 ± 895278.274  ops/s
fourLabels   thrpt   20  75852544.377 ± 972598.185  ops/s
fiveLabels   thrpt   20  58665529.189 ± 443240.792  ops/s
```