@njhill 2019-02-26
For reference I ran the same benchmark on the other proposed impls:

#445:
```
Benchmark     Mode  Cnt         Score        Error  Units
baseline     thrpt   20  90736100.887 ± 297716.947  ops/s
oneLabel     thrpt   20  40977375.778 ± 836712.718  ops/s
twoLabels    thrpt   20  28164696.062 ± 511641.294  ops/s
threeLabels  thrpt   20  24182863.357 ± 175046.611  ops/s
fourLabels   thrpt   20  20844796.684 ± 400396.972  ops/s
fiveLabels   thrpt   20  19438768.621 ± 248130.659  ops/s
```

#459:
```
Benchmark     Mode  Cnt         Score         Error  Units
baseline     thrpt   20  90504144.191 ± 1150495.470  ops/s
oneLabel     thrpt   20  88479586.198 ±  520336.150  ops/s
twoLabels    thrpt   20  65928763.142 ±  464774.439  ops/s
threeLabels  thrpt   20  62229818.280 ±  662305.912  ops/s
fourLabels   thrpt   20  55868210.897 ±  524420.109  ops/s
fiveLabels   thrpt   20  22860553.903 ± 1128285.139  ops/s
```